pyproject.toml
[tool.poetry]
name = "workout_tracker_api"
version = "0.1.0"
description = ""
authors = ["joemoyes <71090896+moyes-joe@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
httpx = "^0.24.1"
mangum = "^0.17.0"
boto3 = "^1.26.137"
mypy-boto3-dynamodb = "^1.26.115"
pyjwt = "^2.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
bandit = "^1.7.5"
moto = "^4.1.10"
pre-commit = "^3.3.2"
pytest-freezegun = "^0.4.2"
mypy = "^1.3.0"
mypy-boto3 = "^1.26.151"

[tool.pytest.ini_options]
testpaths = ["./tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[[tool.mypy.overrides]]
module = [
    "botocore.*",
    "mypy_boto3_dynamodb.*",
    "mangum.*",
    "jwt.*",
    "boto3.*",
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = [".venv", "tests"]

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
